Split out chunks from index
